% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_ancestry_vcfs}
\alias{create_ancestry_vcfs}
\title{Create ancestry-specific VCF files}
\usage{
create_ancestry_vcfs(vcf_path, gt_matrix, pt_matrix, out_african, out_european)
}
\arguments{
\item{vcf_path}{Input VCF path}

\item{gt_matrix}{Genotype matrix}

\item{pt_matrix}{PT (ancestry) matrix}

\item{out_african}{Output path for African-specific VCF}

\item{out_european}{Output path for European-specific VCF}
}
\description{
Create ancestry-specific VCF files
}
