% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry_split.R
\name{read_bed_file}
\alias{read_bed_file}
\title{Read a PLINK .bed file of BED-encoded local ancestry calls}
\usage{
read_bed_file(bed, bim, fam)
}
\arguments{
\item{bed}{Path to the \code{.bed} file.}

\item{bim}{Path to the \code{.bim} file (variant IDs are taken from
column 2).}

\item{fam}{Path to the \code{.fam} file (sample IDs are taken from
column 2, i.e. IID).}
}
\value{
Integer matrix (variants x samples), dimnamed from the
  \code{.bim}/\code{.fam} files. Values: 0 = no call, 1 = homozygous
  first allele, 2 = heterozygous, 3 = homozygous second allele.
}
\description{
Reads a PLINK binary genotype file (\code{.bed}/\code{.bim}/\code{.fam}
trio) into an integer matrix, for the case where local ancestry has been
encoded as PLINK genotype calls at ancestry-tract "SNPs" (as consumed by
\code{src/step1_vcf_split_by_ancestry.R}): homozygous first allele = pure
EUR/EUR (code 1), heterozygous = mixed AFR/EUR (code 2), homozygous
second allele = pure AFR/AFR (code 3), missing = no call (code 0). This
matches \code{snpStats::read.plink()}'s raw numeric convention (values
feed directly into \code{\link{split_diploid}}/\code{\link{ancestry_split_dosage}}
as \code{pt_matrix}/\code{ancestry} without \code{snpStats} as a
dependency.
}
\examples{
ext <- system.file("extdata", package = "lantern")
pt <- read_bed_file(file.path(ext, "toy_ancestry.bed"),
                    file.path(ext, "toy_ancestry.bim"),
                    file.path(ext, "toy_ancestry.fam"))
pt[, 1:6]   # ancestry tracts x samples, codes 1/2/3

}
\seealso{
\code{\link{split_diploid}}, \code{\link{ancestry_split_dosage}}
}
